Added batch generation, fixed prompting
